Split MetabaseAdminPurchase.tsx into two files
